requirements.txt
pandas==3.0.1
numpy==2.3.4
yfinance==1.2.0
google-genai==1.66.0
requests==2.32.5
python-dotenv==1.2.2
feedparser==6.0.12
psutil==5.9.8
groq==1.1.0
